requirements.txt
pykeepass
pyyaml
selenium